new features

+++ src/main.rs
@@ -12,6 +12,12 @@ const BACKUP_FILE_REGEX: &str = "SYNCOPY_BACKUP_(\\d{2}_\\d{2}_\\d{4}_\\d{2}_\\d
 const BACKUP_FILE_PREFIX: &str = "SYNCOPY_BACKUP";
 const DATE_FORMAT: &str = "%d_%m_%Y_%H_%M";
 const CONFIG_FILE_NAME: &str = "config.toml";
+const CLIENT_LINK_PREFIX: &str = "https://disk.yandex.ru/client";
+
+const QUIET_ARG_ID: &str = "quiet";
+const CHECK_ARG_ID: &str = "check";
+const DRY_RUN_ARG_ID: &str = "dry run";
+const DIRTY_ARG_ID: &str = "dirty";
 
 #[derive(Clone, Debug)]
 #[allow(dead_code)]
@@ -34,15 +40,36 @@ async fn main() {
         .author(env!("CARGO_PKG_AUTHORS"))
         .about("Backup utility tethered with Yandex Disk")
         .arg(
-            Arg::with_name("quiet")
+            Arg::with_name(QUIET_ARG_ID)
                 .short("q")
                 .long("quiet")
-                .help("do not output logs")
+                .help("Do not output logs")
+                .required(false),
+        )
+        .arg(
+            Arg::with_name(CHECK_ARG_ID)
+                .short("c")
+                .long("check")
+                .help("Only check previous backups")
+                .required(false),
+        )
+        .arg(
+            Arg::with_name(DRY_RUN_ARG_ID)
+                .short("d")
+                .long("dry-run")
+                .help("Do not upload upon completion")
+                .required(false),
+        )
+        .arg(
+            Arg::with_name(DIRTY_ARG_ID)
+                .short("D")
+                .long("dirty")
+                .help("Do not clean upon completion")
                 .required(false),
         )
         .get_matches();
 
-    let quiet = matches.is_present("quiet");
+    let quiet = matches.is_present(QUIET_ARG_ID);
 
     let logger = tools::Logger { enabled: !quiet };
 
@@ -86,50 +113,56 @@ async fn main() {
         latest_created_at
     ));
 
-    logger.log("Preparing backup...");
+    if !matches.is_present(CHECK_ARG_ID) {
+        logger.log("Preparing backup...");
 
-    let output_file_name = tools::construct_backup_file_name(
-        BACKUP_FILE_PREFIX,
-        &current_date.format(DATE_FORMAT).to_string(),
-    );
-    let output_file = format!("{}{}", config.backups.output_directory, output_file_name);
-    let input_paths = config.backups.include;
+        let output_file_name = tools::construct_backup_file_name(
+            BACKUP_FILE_PREFIX,
+            &current_date.format(DATE_FORMAT).to_string(),
+        );
+        let output_file = format!("{}{}", config.backups.output_directory, output_file_name);
+        let input_paths = config.backups.include;
 
-    let excluded = tools::parse_ignore_string(config.backups.exclude)
-        .unwrap_or_else(|e| panic!("Invalid exclusion config: {}", e));
+        let excluded = tools::parse_ignore_string(config.backups.exclude)
+            .unwrap_or_else(|e| panic!("Invalid exclusion config: {}", e));
 
-    logger.log(&format!("Packing into file '{}'...", output_file));
+        logger.log(&format!("Packing into file '{}'...", output_file));
 
-    let file_size = pack::pack_paths(&input_paths, &output_file, excluded, !quiet)
-        .unwrap_or_else(|e| panic!("Pack failed: {}", e));
+        let file_size = pack::pack_paths(&input_paths, &output_file, excluded, !quiet)
+            .unwrap_or_else(|e| panic!("Pack failed: {}", e));
 
-    logger.log(format!(
-        "  File size: {}",
-        tools::get_bytes_string(file_size)
-    ));
+        logger.log(format!(
+            "  File size: {}",
+            tools::get_bytes_string(file_size)
+        ));
 
-    logger.log("Preparing upload...");
+        if !matches.is_present(DRY_RUN_ARG_ID) {
+            logger.log("Preparing upload...");
 
-    let upload_operation = api.get_upload_operation(&output_file_name).await.unwrap();
-    logger.log(&format!("Uploading to {}...", upload_operation.href));
+            let upload_operation = api.get_upload_operation(&output_file_name).await.unwrap();
+            logger.log(&format!("Uploading to {}...", upload_operation.href));
 
-    let destination = api
-        .upload_file(&upload_operation.href, &output_file)
-        .await
-        .unwrap();
+            let destination = api
+                .upload_file(&upload_operation.href, &output_file)
+                .await
+                .unwrap();
 
-    logger.log(if destination.is_some() {
-        format!(
-            "  Available at https://disk.yandex.ru/client{}",
-            destination.unwrap()
-        )
-    } else {
-        "Upload done".to_string()
-    });
+            logger.log(if destination.is_some() {
+                format!(
+                    "  Available at {}{}",
+                    CLIENT_LINK_PREFIX,
+                    destination.unwrap()
+                )
+            } else {
+                "Upload done".to_string()
+            });
+        }
 
-    logger.log("Cleaning...");
-
-    fs::remove_file(output_file).unwrap_or_else(|e| panic!("Deletion failed: {}", e));
+        if !matches.is_present(DIRTY_ARG_ID) {
+            logger.log("Cleaning...");
+            fs::remove_file(output_file).unwrap_or_else(|e| panic!("Deletion failed: {}", e));
+        }
+    }
 
     logger.log(format!(
         "  Done in {}",
